refactor: move DAT parsers into dat/ subdirectory to mirror dsn/ structure
