+++ src/parsers/cadence/dat/pstxnet-parser.ts
@@ -3,15 +3,15 @@
  * Extracts net connections in the format: { netName: { refdes: [pinNumbers] } }
  */
 
-import { readFile } from 'fs/promises';
-import type { NetConnections } from '../../types.js';
+import { readFile } from "fs/promises";
+import type { NetConnections } from "../../../types.js";
 
 /**
  * Parse a .pstxnet.dat file and extract net connections.
  * Pure function that reads from disk and returns parsed data.
  */
 export const parsePstxnet = async (filePath: string): Promise<NetConnections> => {
-  const content = await readFile(filePath, 'utf-8');
+  const content = await readFile(filePath, "utf-8");
   return parsePstxnetContent(content);
 };
 
@@ -19,7 +19,7 @@ export const parsePstxnet = async (filePath: string): Promise<NetConnections> =>
  * Parse pstxnet file content (pure function for testing).
  */
 export const parsePstxnetContent = (content: string): NetConnections => {
-  const lines = content.split('\n').map((line) => line.trim());
+  const lines = content.split("\n").map((line) => line.trim());
 
   const netConnections: NetConnections = {};
   let currentNet: string | null = null;
@@ -39,13 +39,13 @@ export const parsePstxnetContent = (content: string): NetConnections => {
   };
 
   for (const line of lines) {
-    if (line === 'NET_NAME') {
+    if (line === "NET_NAME") {
       saveCurrentNet();
       currentNet = null;
       currentPins = [];
     } else if (currentNet === null && line.startsWith("'") && line.endsWith("'")) {
       currentNet = line.slice(1, -1);
-    } else if (line.startsWith('NODE_NAME')) {
+    } else if (line.startsWith("NODE_NAME")) {
       const parts = line.split(/\s+/);
       if (parts.length >= 3) {
         const refdes = parts[1];
